+++ rasgoql/rasgoql/data/base.py
@@ -1,11 +1,16 @@
 """
 Base DataWarehouse classes
 """
+from __future__ import annotations
 from abc import ABC
 import re
-from typing import Union
+import os
+from typing import Union, Optional
+from collections import namedtuple
 
 import pandas as pd
+
+FQTN = namedtuple("FQTN", ["database", "schema", "table"], defaults=(None, None, None))
 
 
 class DWCredentials(ABC):
@@ -49,6 +54,30 @@ class DWCredentials(ABC):
         Saves credentials to a .env file on your machine
         """
         raise NotImplementedError()
+
+    @staticmethod
+    def _parse_env_vars(prefix: str) -> dict[str, Union[str, bool, int]]:
+        """
+        Pull out all environment variables starting with a prefix
+        ( usual same name as data warehouse, i.e. REDSHIFT )
+        Auto-detect and convert integers and boolean values.
+        """
+        prefix = prefix.upper()
+        if not prefix.endswith("_"):
+            prefix = f"{prefix}_"
+
+        env_vars = {}
+        for var_name, value in os.environ.items():
+            if var_name.upper().startswith(prefix):
+                # Convert Booleans
+                if value.lower().strip() in ("false", "true"):
+                    value = False if value.lower().strip() == "false" else True
+                # Convert Integers
+                elif value.strip().isnumeric():
+                    value = int(value.strip())
+                env_vars[var_name.lower()[len(prefix):]] = value
+        return env_vars
+
 
 class DataWarehouse(ABC):
     """
@@ -102,23 +131,28 @@ class DataWarehouse(ABC):
     def parse_fqtn(
         self,
         fqtn: str,
-        default_namespace: str = None,
+        default_namespace: Optional[str] = None,
         strict: bool = True
-    ) -> tuple:
+    ) -> FQTN:
         """
         Accepts a possible fully qualified table string and returns its component parts
         """
-        # TODO: review the logic of this method
         if strict:
             fqtn = self.validate_fqtn(fqtn)
-            return (* fqtn.split("."),)
-        database, schema = self.parse_namespace(default_namespace)
+            return FQTN(* fqtn.split("."))
+
         if fqtn.count(".") == 2:
-            return (* fqtn.split("."),)
+            return FQTN(* fqtn.split("."))
+
+        if not default_namespace:
+            raise ValueError(f'{fqtn} is not a well-formed fqtn')
+
+        database, schema = self.parse_namespace(default_namespace)
+
         if fqtn.count(".") == 1:
-            return (database, * fqtn.split("."),)
+            return FQTN(database, * fqtn.split("."))
         if fqtn.count(".") == 0:
-            return (database, schema, fqtn)
+            return FQTN(database, schema, fqtn)
         raise ValueError(f'{fqtn} is not a well-formed fqtn')
 
     def parse_namespace(
@@ -133,7 +167,7 @@ class DataWarehouse(ABC):
 
     def validate_fqtn(self, fqtn: str) -> str:
         """
-        Accepts a possible fully qualified table string and decides whether it is well formed
+        Accepts a possible fully qualified table string and decides whether it is well-formed
         """
         if re.match(r'^[^\s]+\.[^\s]+\.[^\s]+', fqtn):
             return fqtn
@@ -142,9 +176,9 @@ class DataWarehouse(ABC):
     def validate_namespace(
         self,
         namespace: str
-    ) -> bool:
+    ) -> str:
         """
-        Accepts a possible namespace string and decides whether it is well formed
+        Accepts a possible namespace string and decides whether it is well-formed
         """
         if re.match(r'^[^\s]+\.[^\s]+', namespace):
             return namespace