+++ rasgoql/rasgoql/imports.py
@@ -62,8 +62,10 @@ except ImportError:
 try:
     from sqlalchemy import create_engine as alchemy_engine
     from sqlalchemy import exc as alchemy_exceptions
+    from sqlalchemy.engine import URL as alchemy_url
     from sqlalchemy.orm import Session as alchemy_session
 except ImportError:
     alchemy_engine = None
     alchemy_exceptions = None
     alchemy_session = None
+    alchemy_url = None