+++ rasgoql/CHANGELOG.md
@@ -9,7 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Add
 - Add SQLChain to/from YAML workflow
-- Add support for Redshift
 
 ### Change
 - N/A
@@ -76,6 +75,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added failure and execution telemetry for methods in Transform and Dataset classes
 
+## [1.5.0] - 2022-04-11
+### Added
+- Added support for Amazon Redshift
+### Removed
+- Removed `check_data_warehouse` function
+-
 [1.0.0]: https://pypi.org/project/rasgoql/1.0.0/
 [1.0.1]: https://pypi.org/project/rasgoql/1.0.1/
 [1.0.2]: https://pypi.org/project/rasgoql/1.0.2/
@@ -85,3 +90,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.3.0]: https://pypi.org/project/rasgoql/1.3.0/
 [1.3.1]: https://pypi.org/project/rasgoql/1.3.1/
 [1.4.0]: https://pypi.org/project/rasgoql/1.4.0/
+[1.5.0]: https://pypi.org/project/rasgoql/1.5.0/